refactor(icons): update import paths for Eye and EyeOff components
